Move CLI-only byte formatter out of shared utils
